writing task

+++ src/main/java/edu/nc/service/UserService.java
@@ -2,8 +2,12 @@ package edu.nc.service;
 
 import edu.nc.dataaccess.entity.RegUserEntity;
 import edu.nc.dataaccess.entity.User;
+import edu.nc.dataaccess.model.security.Authority;
+import edu.nc.dataaccess.model.security.AuthorityName;
 import edu.nc.dataaccess.model.security.UserEnterWrapper;
+import edu.nc.dataaccess.repository.AuthorityRepository;
 import edu.nc.dataaccess.repository.RegistrationRepository;
+import edu.nc.dataaccess.wrapper.LoginNameLastNameWrapper;
 import edu.nc.dataaccess.wrapper.UpdateUserWrapper;
 import edu.nc.dataaccess.wrapper.registration.LoginAndPassword;
 import edu.nc.dataaccess.wrapper.registration.PersonalData;
@@ -12,6 +16,7 @@ import edu.nc.dataaccess.wrapper.login.TokenWrapper;
 import edu.nc.common.utils.TimeProvider;
 import edu.nc.dataaccess.model.security.UserWrapper;
 import edu.nc.dataaccess.repository.UserRepository;
+import edu.nc.security.JwtUserDetails;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -23,10 +28,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -34,11 +36,13 @@ import java.util.regex.Pattern;
 public class UserService {
     private final UserRepository userRepository;
     private RegistrationRepository registrationRepository;
+    private AuthorityRepository authorityRepository;
 
     private static final String EMAIL_REGULAR_EXPRESSION = "^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
 
     private static final String REGISTRATION_THEME = "Registration";
 
+    private static final String DOMAIN_FRONT = "http://localhost:4200/confirm-registration/";
 
     private static final long DELAY_TO_CONFIRM = 1000 * 3600 * 3;
 
@@ -51,10 +55,12 @@ public class UserService {
     private TimeProvider timeProvider;
 
     @Autowired
-    public UserService(UserRepository userRepository, TimeProvider timeProvider, RegistrationRepository registrationRepository) {
+    public UserService(UserRepository userRepository, TimeProvider timeProvider, RegistrationRepository registrationRepository,
+                       AuthorityRepository authorityRepository) {
         this.userRepository = userRepository;
         this.timeProvider = timeProvider;
         this.registrationRepository = registrationRepository;
+        this.authorityRepository = authorityRepository;
     }
 
     public ResponseEntity<User> getUser(Long userId) {
@@ -65,10 +71,22 @@ public class UserService {
         return new ResponseEntity<>(userRepository.findAll(new Sort("username")), HttpStatus.OK);
     }
 
-    public ResponseEntity<User> createUser(UserEnterWrapper userWrapper) {
-        User newUser = new User(userWrapper);
+    public ResponseEntity<User> createUser(UserWrapper userWrapper) {
+        Optional<Authority> optionalAuthority = authorityRepository.findByName(AuthorityName.ROLE_USER);
+        Authority currentAuthority;
+        if(!optionalAuthority.isPresent()){
+            Authority authority = new Authority();
+            authority.setName(AuthorityName.ROLE_USER);
+            authority.setUsers(new ArrayList<>());
+            currentAuthority = authorityRepository.saveAndFlush(authority);
+        }else {
+            currentAuthority = optionalAuthority.get();
+        }
+        User newUser = new User(userWrapper, currentAuthority);
         newUser = userRepository.save(newUser);
-        return new ResponseEntity<>(newUser, HttpStatus.CREATED);
+
+
+        return new ResponseEntity<>( HttpStatus.CREATED);
     }
 
     public ResponseEntity<TokenWrapper> authenticateUser(UserEnterWrapper userEnterWrapper,
@@ -126,14 +144,20 @@ public class UserService {
         }
         //find the same entry
         Optional<RegUserEntity> optionalRegUserEntity = registrationRepository.findByLogin(wrapper.getLogin());
+        //find this email in user DB
+        User user = userRepository.findByUsername(wrapper.getLogin());
+        if (user != null) {
+            return new ResponseEntity(HttpStatus.BAD_REQUEST);
+        }
         //delete the entry
         optionalRegUserEntity.ifPresent(regUserEntity -> registrationRepository.delete(regUserEntity.getId()));
+
 
         RegUserEntity entity = new RegUserEntity(wrapper.getLogin(), wrapper.getPassword());
         String secret = entity.getUuid();
         registrationRepository.saveAndFlush(entity);
 
-        EmailSender sender = new EmailSender(wrapper.getLogin(), REGISTRATION_THEME, secret);
+        EmailSender sender = new EmailSender(wrapper.getLogin(), REGISTRATION_THEME, DOMAIN_FRONT + secret);
         sender.send();
 
         return new ResponseEntity(HttpStatus.OK);
@@ -149,17 +173,32 @@ public class UserService {
             return new ResponseEntity(HttpStatus.REQUEST_TIMEOUT);
         }
         registrationRepository.delete(current.getId());
-//        return this.createUser(new UserWrapper( current.getLogin(),
-//                                                current.getPasswordHash(),
-//                                                wrapper.getUserFirstname(),
-//                                                wrapper.getUserLastname(),
-//                                                wrapper.getUserDateOfBirth(),
-//                                                0,
-//                                                true));
-        return null;
+        return this.createUser(new UserWrapper( current.getLogin(),
+                current.getPasswordHash(),
+                wrapper.getUserFirstname(),
+                wrapper.getUserLastname(),
+                0));
     }
 
     private Date calculateExpirationDate(Date createdDate) {
         return new Date(createdDate.getTime() + expiration * 1000);
+    }
+
+    public ResponseEntity<LoginNameLastNameWrapper> getUserData(){
+        User current = userRepository.findByUsername(JwtUserDetails.getUserName());
+        return new ResponseEntity<>(new LoginNameLastNameWrapper(current.getUsername(),
+                current.getFirstname(), current.getLastname()), HttpStatus.OK);
+    }
+
+
+    public ResponseEntity isAdmin(){
+        User current = userRepository.findByUsername(JwtUserDetails.getUserName());
+        List<Authority> authorities = current.getAuthorities();
+        for (Authority x : authorities) {
+            if(x.getName() == AuthorityName.ROLE_ADMIN){
+                return new ResponseEntity<>(Boolean.TRUE, HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(Boolean.FALSE, HttpStatus.OK);
     }
 }
